requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-core>=2.14.0
typing-extensions>=4.8.0

# AI/ML dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
langchain-community>=0.0.20
langchain-perplexity>=0.1.0
langchain_tavily>=0.1.0
langchain-text-splitters>=0.0.1
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
lxml>=4.9.0
tavily-python>=0.3.0
html2text>=2020.1.16
beautifulsoup4>=4.12.0
azure-messaging-webpubsubservice>=1.0.0
aiohttp>=3.9.0


